+++ examples/api/dynamic-form-element-options/src/yes-no-maybe.js
@@ -5,17 +5,17 @@ module.exports.get = function () {
     {
       value: 'y',
       label: 'Yes',
-      colour: '#4CAF50' // Green
+      colour: '#4CAF50', // Green
     },
     {
       value: 'n',
       label: 'No',
-      colour: '#F44336' // Red
+      colour: '#F44336', // Red
     },
     {
       value: 'm',
       label: 'Maybe',
-      colour: '#2196F3' // Blue
-    }
+      colour: '#2196F3', // Blue
+    },
   ]
 }