+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **[BREAKING]** `request.url.query` to support multi-value query string parameters. This means the object may contain entries that have a `string` value **or** a `string[]` value
+
+### Added
+
+- `request.url.querystring` to API Hosting request object
+
 ## [0.2.4] - 2022-01-19
 
 ### Fixed